% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dispersion.R
\name{beta.dispersion}
\alias{beta.dispersion}
\title{Estimate Beta-dispersion}
\usage{
beta.dispersion(
  mD,
  group,
  metrics = NULL,
  type = c("median", "centroid"),
  bias.adjust = FALSE,
  sqrt.dist = FALSE,
  add = FALSE,
  update.mD = TRUE
)
}
\arguments{
\item{mD}{required; the `microbData` object, with distance matrices (see \code{\link[microbData]{beta.diversity}}) from which to estimate beta-dispersion.}

\item{group}{required; a character string indicating by which column in the Metadata table the samples should be grouped.}

\item{metrics}{character; if NULL, beta-dispersion will be calculated for each distance matrix present in the `microbData` object. If metrics are provided, beta-dispersion will only be calculated for those metrics. Default is NULL.}

\item{type}{character; the type of analysis to perform. Use the spatial median ("median") or the group centroid ("centroid")? Default is "median". (Directly from \code{\link[vegan]{betadisper}}).}

\item{bias.adjust}{logical: adjust for small sample bias in beta diversity estimates? Default is FALSE.}

\item{sqrt.dist}{logical; take square root of dissimilarities. This often euclidifies dissimilarities. Default is FALSE. (Directly from \code{\link[vegan]{betadisper}}).}

\item{add}{logical; add a constant to the non-diagonal dissimilarities such that all eigenvalues are non-negative in the underlying Principal Co-ordinates Analysis (see \code{\linkl[vegan]{wcmdscale}} for details). Choice "lingoes" (or TRUE) use the recommended method of Legendre & Anderson (1999: “method 1”) and "cailliez" uses their “method 2”. Default is FALSE. (Directly from \code{\link[vegan]{betadisper}}).}

\item{update.mD}{logical; should this function return a new microbData object with the beta-dispersion results added to the Metadata table and add `metrics` to Other.data (TRUE) or just the results of the beta-dispersion estimation (FALSE)? Default is TRUE.}
}
\value{
If `update.mD` is TRUE, a `microbData` object with estimated beta-dispersion scores added to the Metadata table and a list of the "betadisper" objects added to the Other.data slot.

If `update.mD` is FALSE, a list containing:

\item{Distances}{a table of the estimated beta-dispersion scores for each sample}
\item{Betadispers}{a list of the "betadisper" objects}
}
\description{
Utilizes \code{\link[vegan]{betadisper}} to calculate beta-dispersion (distance to centroid) for each sample by a group.
}
\details{
This function utilizes \code{\link[vegan]{betadisper}} to estimate beta-dispersion (i.e., distance from spatial median or group centroid in ordination space) by categorical group. See `?vegan::betadisper` for more details about the procedure that function employs.
}
\seealso{
\code{\link[vegan]{betadisper}}
}
