% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbData.R
\name{get.microbData}
\alias{get.microbData}
\alias{get.abundances}
\alias{get.distance.matrices}
\alias{get.assignments}
\alias{get.feature.col}
\alias{get.feature.names}
\alias{get.metadata}
\alias{get.other.data}
\alias{get.phylogeny}
\alias{get.sample.col}
\alias{get.sample.names}
\alias{get.covariate.names}
\title{Get Abundances}
\usage{
get.microbData(mD, slot.name, return.as = return.as.choices)

get.abundances(mD, return.as = return.as.choices)

get.distance.matrices(mD)

get.assignments(mD, return.as = return.as.choices)

get.feature.col(mD)

get.feature.names(mD)

get.metadata(mD, return.as = return.as.choices)

get.other.data(mD, location)

get.phylogeny(mD)

get.sample.col(mD)

get.sample.names(mD)

get.covariate.names(mD)
}
\description{
Get the abundance table from a \code{microbData} object.

Get the abundance table from a \code{microbData} object.

Get the distance matrices from a \code{microbData} object.

Get the feature assignments table from a \code{microbData} object.

Get the name of the column containing specific feature IDs from the features table in a \code{microbData} object.

Get the specific IDs of features in a \code{microbData} object.

Get the metadata table from a \code{microbData} object.

Get an element from the Other.data list of a \code{microbData} object.

Get the phylogenetic tree from a \code{microbData} object.

Get the name of the column containing specific sample IDs from the samples table in a \code{microbData} object.

Get the specific IDs of samples in a \code{microbData} object.

Get the column names from the Metadata (minus sample column name) in a \code{microbData} object.
}
