% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glom_features.R
\name{glom.features}
\alias{glom.features}
\title{Agglomerate Feature Abundances by Level}
\usage{
glom.features(mD, level, decreasing.assignments = TRUE)
}
\arguments{
\item{mD}{required; the \code{microbData} object to be glommed. It must contain a Assignments table.}

\item{level}{required; a character matching a column name in the Assignments table by which to glom (sum) abundances.}

\item{decreasing.assignments}{logical or NULL; in the Assignments table, are assignments decreasing in hierarchy (e.g. Phylum down to Genus), or increasing (e.g. KO up to Pathway). If you do not want the returning Feature table trimmed of lower level assignments, set to NULL. Default is TRUE.}
}
\value{
a \code{microbData} object with a new Abundances and Feature table that reflect the glomming process.
}
\description{
Sum feature counts by the specified assignment level per sample. E.g., sum ASVs by their Genus assignment.
}
\details{
This function will sum feature abundances for any assignment level that is present in the Assignments table. Note the \code{decreasing.assignments} argument. It is assumed that the Assignments table will be organized in some sort of hierarchy, and will only keep the columns of that feature table that are the same level as an higher than the column passed to \code{level}.

Also note that this function will drop any phylogenetic tree that was present as it cannot accurately reflect the phylogeny of the glommed features. Also note that things like count transformation and alpha-, beta-diversity, and beta-dispersion results already added to Metadata, or notes about them in Other.data will remain there, even if no longer accurate. Therefore it is important to run this function on a \code{microbData} object with raw counts and no other analyses conducted on it, beyond, perhaps, filtering.
}
