% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_seqs.R
\name{add.guide.seqs}
\alias{add.guide.seqs}
\alias{remove.guide.seqs}
\title{Dealing with Guide Seqs}
\usage{
add.guide.seqs(
  asv.seqs.file,
  guide.seqs.file,
  joined.file,
  keep.guide.ids = TRUE
)

remove.guide.seqs(guide.seq.ids, full.tree.file, asvs.only.tree.file)
}
\arguments{
\item{asv.seqs.file}{required; FASTA file containing ASV sequences.}

\item{guide.seqs.file}{required; FASTA file containing guide sequences.}

\item{joined.file}{required; FASTA file to write concatenated sequences to.}

\item{keep.guide.ids}{logical; whether to return a vector containing guide sequence IDs for later removal from output. Default is TRUE}

\item{guide.seq.ids}{required; a charater vector of the guide sequence IDs, usually from running `add.guide.seqs`.}

\item{full.tree.file}{required; the file containing the phylogenetic tree containing both ASV and guide sequence IDs in Newick format.}

\item{asvs.only.tree.file}{required; the file to write the pruned tree of just ASV IDs to. Will be written in Newick format.}
}
\description{
Functions for add and remove guide seqs from FASTAs, alignments, and trees.
}
\seealso{
\code{\link{generate.tool.command}}, \code{\link{system}}
}
