% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mD.raw}
\alias{mD.raw}
\title{Example Unrarefied microbData}
\format{
A microbData object with unrarefied counts
}
\usage{
mD.raw
}
\description{
Example Unrarefied microbData
}
\keyword{datasets}
