% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.microbData.R
\name{merge.microbData}
\alias{merge.microbData}
\title{Merge Two MicrobDatas}
\usage{
\method{merge}{microbData}(mD1, mD2)
}
\arguments{
\item{mD1}{required; a \code{microbData} object you want to merge.}

\item{mD2}{required; a \code{microbData} object you want to merge.}
}
\description{
Merge two \code{microbData} objects.
}
\seealso{
\code{\link[microbData]{microbData}}
}
