% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_microbData.R
\name{merge_mDs}
\alias{merge_mDs}
\title{Merge Two MicrobDatas}
\usage{
merge_mDs(mD1, mD2)
}
\arguments{
\item{mD1}{required; a \code{microbData} object you want to merge.}

\item{mD2}{required; a \code{microbData} object you want to merge.}
}
\description{
Merge two \code{microbData} objects.
}
\seealso{
\code{\link[microbData]{microbData}}
}
