% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbData2phyloseq.R
\name{microbData2phyloseq}
\alias{microbData2phyloseq}
\title{MicrobData to Phyloseq}
\usage{
microbData2phyloseq(mD)
}
\arguments{
\item{mD}{required; the microbData object to be converted to a phyloseq object.}
}
\description{
Convert a microbData object to a phyloseq object
}
