% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothur_NAST_align.R
\name{mothur.NAST.align}
\alias{mothur.NAST.align}
\title{NAST Alignment with `mothur`}
\usage{
mothur.NAST.align(
  fasta,
  template.file,
  mothur.path = NULL,
  output.dir = getwd(),
  ncores = 1
)
}
\arguments{
\item{fasta}{required; the path and name of the FASTA file containing the sequences to be aligned}

\item{template.file}{required; the path and name of the alignment file to be used. See https://mothur.org/wiki/alignment_database/ to download a database of your choice.}

\item{mothur.path}{character; path to `mothur` executable if it is not in your system PATH variable. Default is NULL.}

\item{output.dir}{character; path to output directory. Default is `getwd()`.}

\item{ncores}{integer; the number of cores for `mothur`.}
}
\description{
Uses the `align.seqs` function from `mothur` (requires install: https://mothur.org/wiki/installation/)
}
\seealso{
\code{\link{generate.tool.command}}, \code{\link{system}}
}
