% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate.R
\name{ordinate}
\alias{ordinate}
\title{Create an Ordination}
\usage{
ordinate(
  mD,
  method = "dbRDA",
  formula = NULL,
  only.for = NULL,
  include.feature.scores = FALSE,
  update.mD = TRUE,
  ...
)
}
\arguments{
\item{mD}{required; microbData object from which data will be read.}

\item{method}{character; the name of the ordination method to be used. Currently supported methods are "PCoA" (principal coordinate analysis; \code{\link[ape]{pcoa}}), "NMDS" (nonmetric multidimensional scaling; \code{\link[vegan]{metaMDS}}), and "dbRDA" (distance-based redundancy analysis; \code{\link[]{capscale}}). The last requires a formula, as it is a constrained ordination method. Default is "dbRDA".}

\item{formula}{formula; in the form of e.g., `~ Var1 + Var2 * Var3`. This is only required for and used by the dbRDA method. Default is NULL.}

\item{only.for}{character or integer; a vector of the names or the indices of the matrices contained in the Distance.matrices slot of the microbData. If NULL, will make an ordination for all matrices. Default is NULL.}

\item{include.feature.scores}{logical; some ordination methods, like "NMDS" and "dbRDA" allow feature (often referred to a 'species') scores as well as site scores. If true, these will be calculated. Default is FALSE.}

\item{update.mD}{logical; should this function return a new microbData object with the ordinations list (called "Ordinations") added to Other.data (TRUE) or just a list of the results (FALSE)? Default is TRUE.}

\item{...}{other arguments passed to ordination methods. See their documentation for appropriate arguments.}
}
\description{
Create an ordination from a microbData object. The microbData must have Distance Matrices (can be created with \code{\link[microbData]{beta.diversity}}), and an ordination will be created for each matrix unless otherwise specified.
}
\seealso{
\code{\link[ape]{pcoa}}, \code{\link[vegan]{metaMDS}}, \code{\link[vegan]{capscale}}, \code{\link[ape]{pcoa}},  \code{\link[vegan]{sppscores}}
}
