% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination_coords.R
\name{ordination.coords}
\alias{ordination.coords}
\title{Get Plotting Coordinates from (an) Ordination(s)}
\usage{
ordination.coords(
  mD = NULL,
  ord.list.name = "Ordinations",
  ord.list = NULL,
  metadata = NULL,
  feature.coords = FALSE,
  feature.tbl = NULL,
  constraint.coords = FALSE,
  combine = TRUE,
  axes = 1:2,
  axis.digits = 2
)
}
\arguments{
\item{mD}{microbData; a microbData object that includes ordination results (typically labelled "Ordinations" in the Other.data slot). Required if `ord.list` and `metadata` are NULL. Default is NULL.}

\item{ord.list.name}{character; required if `mD` has an list of ordinations stores under something other than Other.data$Ordinations. Default is "Ordinations".}

\item{ord.list}{list; a list of ordinations, required if `mD` is NULL. This list **must** be named if `combine` is TRUE. Default is NULL.}

\item{metadata}{array; a data.frame/data.table/matrix with metadata for the samples used in the ordination(s). Default is NULL.}

\item{feature.coords}{logical; if FALSE will only get coordinates for samples. If TRUE, will also get coordinates for features (e.g. ASVs, KOs, ...). Default is FALSE.
#' @param feature.tbl array; a data.frame/data.table/matrix with data for the features used in the ordination(s). This is only necessary if there is no `mD` and `feature.coords` is TRUE. Default is NULL.}

\item{constraint.coords}{logical; only relevant for dbRDA ordinations. If FALSE, will only get coordinates for samples. If TRUE, will also get coordinates for constraining variables to plot as vectors. Default is FALSE.}

\item{combine}{logical; if TRUE, will combine related results into data.tables and add columns of ordination names so results can all be plotted together, e.g. by using the \code{\link[ggplot2]{facet_wrap}} or \code{\link[ggplot2]{facet_grid}} functions with \code{\link[ggplot2]{ggplot}}. If FALSE, will return a list of data.tables with coordinates. Default is TRUE.}
}
\description{
Get the appropriate coordinates to plot the results of one or more ordinations. This also returns axis labels with percentage variance explained for relevant ordination methods (i.e., for all supported methods but NMDS).
}
\seealso{
\code{\link{}}
}
