% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_NA_features.R
\name{rename.NA.features}
\alias{rename.NA.features}
\title{Rename NA Feature Assignments}
\usage{
rename.NA.features(mD, force.split = FALSE, level.order = NULL)
}
\arguments{
\item{mD}{required; microbData object to read from and write to.}

\item{force.split}{logical; should the lowest feature assignment (e.g. ASV) be appended to replacement assignment? If TRUE, this will prevent features that have NAs at the same assignment level from receiving identical assignments (keeping separate during future glomming). If FALSE, NAs will be replaced by a string including the previous level assignment and the current assignment level. Default is FALSE.}

\item{level.order}{character; a vector specifying the hierarchical ordering of the assignment levels. Must be identical to, or a subset of the column names in the Feature Assignments table. If NULL, the order will be inferred as Feature column lowest and the rest in order as the table columns. Default is NULL.}
}
\description{
This function searches through the Features table of a microbData object and replaces any NA assignments with a more descriptive term based on the higher order assignment provided. This prevents unrelated features lacking labels from getting grouped together simply because they're labelled "NA".
}
