% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FastTree2.R
\name{run.FastTree2}
\alias{run.FastTree2}
\title{Build a Phylogenetic Tree with FastTree2}
\usage{
run.FastTree2(
  aligned.seqs.file,
  output.file,
  fasttree.path = NULL,
  log.file = NULL,
  ncores = 1
)
}
\arguments{
\item{aligned.seqs.file}{required; FASTA file containing aligned sequences to build tree with.}

\item{output.file}{required; file name to write output, should have the '.nwk' extension. If not, it will be added.}

\item{fasttree.path}{character; path to \code{mothur} executable if it is not in your system PATH variable. Default is NULL.}
}
\description{
Build an approximately-maximum-likelihood with FastTree2 (http://www.microbesonline.org/fasttree/)
}
\seealso{
\code{\link[microbData]{generate.tool.command}}, \code{\link{system}}
}
