% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{show-microbData}
\alias{show-microbData}
\title{Method extensions to `show` for microbData object}
\usage{
\S4method{show}{microbData}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Method extensions to `show` for microbData object
}
\seealso{
\code{\link[methods]{show}}
}
