% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy.dt}
\alias{taxonomy.dt}
\title{Example Taxonomic Assignments}
\format{
A data.table with example taxonomicy assignments for the ASVs
}
\usage{
taxonomy.dt
}
\description{
Example Taxonomic Assignments
}
\keyword{datasets}
