% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_counts.R
\name{rarefy}
\alias{rarefy}
\alias{center.log.ratio}
\alias{variance.stabilize}
\alias{relative.abundance}
\title{Rarefy Feature Counts}
\usage{
rarefy(
  mD,
  min.abund = 10000,
  exactly.to = NULL,
  trim.features = TRUE,
  user.seed = NULL,
  update.mD = TRUE,
  quiet = FALSE
)

center.log.ratio(
  mD,
  min.abund = 10000,
  min.prop = 0.001,
  min.occur = 0,
  smpls.by.row = TRUE,
  method = "CZM",
  lab = 0,
  update.mD = TRUE,
  quiet = FALSE
)

variance.stabilize(mD, design, update.mD = TRUE)

relative.abundance(mD, update.mD = TRUE)
}
\description{
Rarefy feature counts in a \code{microbData} object to the same depth.

Transform feature counts using the center log-ratio method. This function requires packages [CoDaSeq](https://github.com/ggloor/CoDaSeq) and \code{\link{zCompositions}}.

Transform feature counts using the variance-stabilized method. This function requires package \code{\link{DESeq2}}.

Transform feature counts to relative abundance. This function is the equivalent of \code{transform.counts(mD = mD, f = function(x) x / sum(x))}, but because it's a popular method of transformation, it has its own wrapper here.
}
