% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_feature_names.R
\name{update.feature.names}
\alias{update.feature.names}
\title{Update Feature Names}
\usage{
\method{update}{feature.names}(mD, new.names)
}
\arguments{
\item{mD}{required; microbData object to read from and write to.}

\item{new.names}{required; a character vector of new feature names. This vector must be the same length as the original Feature Names and in the same order.}
}
\description{
A simple way to replace the feature names in a microbData object.
}
