% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_microbData.R
\name{replace.metadata}
\alias{replace.metadata}
\alias{replace.abundances}
\alias{add.distance.matrices}
\alias{add.assignments}
\alias{replace.assignments}
\alias{add.other.data}
\alias{add.phylogeny}
\alias{replace.phylogeny}
\title{Add or replace data in appropriate slot of a \code{microbData} object}
\usage{
replace.metadata(mD, new.tbl)

replace.abundances(mD, new.tbl)

add.distance.matrices(mD, distance.matrices)

add.assignments(mD, assignments, feature.names = NULL)

add.phylogeny(mD, phylo)

add.other.data(mD, name, x, replace = FALSE)
}
\arguments{
\item{mD}{required; the \code{microbData} object to be updated.}

\item{new.tbl}{required for \code{replace.metadata} and \code{replace.abundances}; a data.table or matrix to replace the current Metadata or Abundances slot. Must have the same samples and/or features names as table it replaces.}

\item{distance.matrices}{required for \code{add.distance.matrices}; a single distance matrix  of class "dist" or a list of distance matrices of class "dist".}

\item{assignments}{required for \code{add.assignments}; data.table, data.frame, or matrix table containing taxonomic or functional assignments for each feature (taxon/function). If not already a data.table, the data will be converted to a data.table and the row names will be saved under a column called "Feature". If a keyed data.table (see \code{\link[data.table]{setkey}}), \code{feature.names} will be set from the keyed column.}

\item{phylo}{required for \code{add.phylogeny}; phylo, a phylogenetic tree object.}

\item{name}{required for \code{add.other.data}; the name to give the object in the Other.data list (e.g., "Ordinations" or "Beta.metrics").}

\item{x}{required; a list, vector, or array (data.frame, data.table, matrix, ...) to add to \code{microbData} object in the appropriate (usually Other.data) slot.}

\item{replace}{logical; for \code{add.other.data}. If TRUE, the new data will replace any data currently in that element of the Other.data list. If FALSE, the new data will be appended to any data currently in that element of the Other.data list. Default is FALSE.}
}
\value{
A \code{microbData} object with the new data.
}
\description{
Functions for adding/replacing metadata, abundances, features, phylogeny, distance matrices, and other data to an established \code{microbData} object.

Replace the data.table in the Metadata slot of a \code{microbData} object with a new one

Replace the matrix in the Abundances slot of a \code{microbData} object with a new one

Add one or a list of distance matrices to an already created \code{microbData} object

Add a assignments table, or replace existing assignments table in an already created \code{microbData} object. An alias for this function is \code{replace.assignments}.

Add a phylogenetic to an already created \code{microbData} object

Add a vector, list, etc... to an already created \code{microbData} object in the Other.data slot
}
\details{
These function can replace data in or add data to an existing \code{microbData} object. The \code{@} operator, or the \code{\link{slot}} function can also accomplish this same task. However, for replacing metadata, and abundances, these functions check to make sure the replacement tables match the samples and/or features of the original tables, helping to safeguard against user error. Additionally, these functions are more user friendly and make sure data get into the correct slot than using the base functions.
}
\seealso{
\code{\link[microbData]{microbData}}, \code{\link{slot}}
}
