% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbData.R
\name{get.microbData}
\alias{get.microbData}
\alias{get.abundances}
\alias{get.distance.matrices}
\alias{get.features}
\alias{get.feature.col}
\alias{get.feature.names}
\alias{get.metadata}
\alias{get.other.data}
\alias{get.phylogeny}
\alias{get.sample.col}
\alias{get.sample.names}
\alias{get.covariate.names}
\title{Get data from appropriate slot of a microbData object}
\usage{
get.microbData(mD, slot.name, as.DT = FALSE)

get.abundances(mD, as.DT = FALSE)

get.distance.matrices(mD)

get.features(mD)

get.feature.col(mD)

get.feature.names(mD)

get.metadata(mD)

get.other.data(mD, location)

get.phylogeny(mD)

get.sample.col(mD)

get.sample.names(mD)

get.covariate.names(mD)
}
\arguments{
\item{mD}{required; the microbData object to get data from.}

\item{slot.name}{required for `get.microbData`; the name of the slot (e.g., Metadata, Abundances, ...) you wish to extract from the `mD`. To see what slot names are valid, use the function \code{\link{slotNames}} on the mD.}

\item{as.DT}{logical; for `get.microbData` and `get.abundances`. The Abundances table is the only one stored as a matrix in the microbData. However, if you would like to retrieve it and convert it to a data.table, you can set this to TRUE. Default is FALSE.}

\item{location}{required for `get.other.data`; the index or name of the list element to get from the Other.data slot.}
}
\description{
Functions grabbing features, phylogeny, distance matrices, or other data from a microbData object. The function `get.microbData` allows the user to programmatically get data from various slots as it takes the slot name as a character and returns the data in that slot. The other functions are very simple wrappers for the `@` extraction operator to make code more human-readable, or just because they may be easier to understand.

Get the abundance table from a microbData object.

Get the abundance table from a microbData object.

Get the distance matrices from a microbData object.

Get the features table from a microbData object.

Get the name of the column containing specific feature IDs from the features table in a microbData object.

Get the specific IDs of features in a microbData object.

Get the metadata table from a microbData object.

Get an element from the Other.data list of a microbData object.

Get the phylogenetic tree from a microbData object.

Get the name of the column containing specific sample IDs from the samples table in a microbData object.

Get the specific IDs of samples in a microbData object.

Get the column names from the Metadata (minus sample column name) in a microbData object.
}
\details{
There are two ways to access slots in an object that has them. One can use the `slot` function or the `@` operator. These functions provide wrappers for these two basic functions that make code a bit more human readable. The `get.microbData` function is included in case users want to programmatically grab data from a slot (i.e., store the slot name in a variable and use that to retrieve the data). These functions also have an option to coerce the results in to a `data.table` if it is not already stored as that class.

The rest of the functions are equivalent to using the `@` operator. For example `get.metadata(mD)` is the same as `mD@Metadata`. Neither should cause the user any problem, it is just a matter of preference.

The `get.covariate.names` function is a little bit different in that it doesn't grab an object from a slot, but rather returns the column names from the Metadata table (excluding the sample IDs column).
}
\seealso{
\code{\link{slot}}, \code{\link{\@}}
}
