% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_features.R
\name{keep.features}
\alias{keep.features}
\alias{drop.features}
\alias{filter.features}
\title{Keep, Drop, or Filter Features in a microbData object}
\usage{
keep.features(mD, features)

drop.features(mD, features)

filter.features(mD, ...)
}
\arguments{
\item{mD}{required; the microbData object to be updated.}

\item{features}{character or logical; a vector of feature names to keep or drop. Also can be a logical vector that is either named or in the exact same order as those in the Feature Names slot.}

\item{...}{logical expression(s) by which to filter features based on the Features table, e.g. `Kingdom == "Bacteria" & Class != "Mitochondria"`.}
}
\description{
Functions to specify specific features to keep/drop by name or to filter features based on variable(s) in the Metadata

Create a new microbData objects with just the specified features

Create a new microbData objects without the specified features

Create a new microbData objects with features that match the filter
}
\seealso{
\code{\link[microbData]{microbData}}
}
